test: better output on cli runner
