+++ frappe/commands/utils.py
@@ -809,7 +809,7 @@ def run_tests(
 			click.secho(f"bench --site {site} set-config allow_tests true", fg="green")
 			return
 
-		ret = frappe.test_runner.main(
+		unit_ret, integration_ret = frappe.test_runner.main(
 			site,
 			app,
 			module,
@@ -828,9 +828,15 @@ def run_tests(
 			pdb_on_exceptions=pdb_on_exceptions,
 		)
 
-		if len(ret.failures) == 0 and len(ret.errors) == 0:
+		if (
+			len(unit_ret.failures) == 0
+			and len(unit_ret.errors) == 0
+			and len(integration_ret.failures) == 0
+			and len(integration_ret.errors) == 0
+		):
 			ret = 0
-
+		else:
+			ret = (unit_ret, integration_ret)
 		if os.environ.get("CI"):
 			sys.exit(ret)
