+++ frappe/parallel_test_runner.py
@@ -12,7 +12,7 @@ import requests
 
 import frappe
 
-from .test_runner import SLOW_TEST_THRESHOLD, make_test_records
+from .test_runner import TestResult, make_test_records
 
 click_ctx = click.get_current_context(True)
 if click_ctx:
@@ -70,7 +70,7 @@ class ParallelTestRunner:
 		click.echo(f"Estimated total tests for build {self.build_number}: {self.total_tests}")
 
 	def run_tests(self):
-		self.test_result = ParallelTestResult(stream=sys.stderr, descriptions=True, verbosity=2)
+		self.test_result = TestResult(stream=sys.stderr, descriptions=True, verbosity=2)
 
 		for test_file_info in self.test_file_list:
 			self.run_tests_for_file(test_file_info)
@@ -173,62 +173,6 @@ def split_by_weight(work, weights, chunk_count):
 	assert len(chunks) == chunk_count
 
 	return chunks
-
-
-class ParallelTestResult(unittest.TextTestResult):
-	def startTest(self, test):
-		self.tb_locals = True
-		self._started_at = time.monotonic()
-		super(unittest.TextTestResult, self).startTest(test)
-		test_class = unittest.util.strclass(test.__class__)
-		if not hasattr(self, "current_test_class") or self.current_test_class != test_class:
-			click.echo(f"\n{unittest.util.strclass(test.__class__)}")
-			self.current_test_class = test_class
-
-	def getTestMethodName(self, test):
-		return test._testMethodName if hasattr(test, "_testMethodName") else str(test)
-
-	def addSuccess(self, test):
-		super(unittest.TextTestResult, self).addSuccess(test)
-		elapsed = time.monotonic() - self._started_at
-		threshold_passed = elapsed >= SLOW_TEST_THRESHOLD
-		elapsed = click.style(f" ({elapsed:.03}s)", fg="red") if threshold_passed else ""
-		click.echo(f"  {click.style(' ✔ ', fg='green')} {self.getTestMethodName(test)}{elapsed}")
-
-	def addError(self, test, err):
-		super(unittest.TextTestResult, self).addError(test, err)
-		click.echo(f"  {click.style(' ✖ ', fg='red')} {self.getTestMethodName(test)}")
-
-	def addFailure(self, test, err):
-		super(unittest.TextTestResult, self).addFailure(test, err)
-		click.echo(f"  {click.style(' ✖ ', fg='red')} {self.getTestMethodName(test)}")
-
-	def addSkip(self, test, reason):
-		super(unittest.TextTestResult, self).addSkip(test, reason)
-		click.echo(f"  {click.style(' = ', fg='white')} {self.getTestMethodName(test)}")
-
-	def addExpectedFailure(self, test, err):
-		super(unittest.TextTestResult, self).addExpectedFailure(test, err)
-		click.echo(f"  {click.style(' ✖ ', fg='red')} {self.getTestMethodName(test)}")
-
-	def addUnexpectedSuccess(self, test):
-		super(unittest.TextTestResult, self).addUnexpectedSuccess(test)
-		click.echo(f"  {click.style(' ✔ ', fg='green')} {self.getTestMethodName(test)}")
-
-	def printErrors(self):
-		click.echo("\n")
-		self.printErrorList(" ERROR ", self.errors, "red")
-		self.printErrorList(" FAIL ", self.failures, "red")
-
-	def printErrorList(self, flavour, errors, color):
-		for test, err in errors:
-			click.echo(self.separator1)
-			click.echo(f"{click.style(flavour, bg=color)} {self.getDescription(test)}")
-			click.echo(self.separator2)
-			click.echo(err)
-
-	def __str__(self):
-		return f"Tests: {self.testsRun}, Failing: {len(self.failures)}, Errors: {len(self.errors)}"
 
 
 def get_all_tests(app):
